+++ src/bloqade/builder/assign.py
@@ -69,7 +69,7 @@ class Assign(BatchAssignable, AddArgs, Parallelizable, BackendRoute, AssignBase)
     def __init__(
         self, assignments: Dict[str, ParamType], parent: Optional[Builder] = None
     ) -> None:
-        from bloqade.analysis.common.scan_variables import ScanVariables
+        from bloqade.compiler.analysis.common.scan_variables import ScanVariables
 
         super().__init__(parent)
 
@@ -85,7 +85,7 @@ class BatchAssign(AddArgs, Parallelizable, BackendRoute, AssignBase):
     def __init__(
         self, assignments: Dict[str, List[ParamType]], parent: Optional[Builder] = None
     ) -> None:
-        from bloqade.analysis.common.scan_variables import ScanVariables
+        from bloqade.compiler.analysis.common.scan_variables import ScanVariables
 
         super().__init__(parent)
 
@@ -120,7 +120,7 @@ class ListAssign(AddArgs, Parallelizable, BackendRoute, AssignBase):
         batch_params: Sequence[Dict[str, ParamType]],
         parent: Optional[Builder] = None,
     ) -> None:
-        from bloqade.analysis.common.scan_variables import ScanVariables
+        from bloqade.compiler.analysis.common.scan_variables import ScanVariables
 
         super().__init__(parent)
