+++ src/bloqade/compiler/analysis/common/is_constant.py
