+++ src/bloqade/compiler/analysis/common/is_hyperfine.py
