+++ src/bloqade/compiler/analysis/common/scan_channels.py
