+++ src/bloqade/compiler/analysis/common/scan_variables.py
