+++ src/bloqade/compiler/analysis/hardware/channels.py
