+++ src/bloqade/compiler/analysis/hardware/piecewise_constant.py
