+++ src/bloqade/compiler/passes/hardware/components.py
@@ -1,6 +1,6 @@
-from bloqade.codegen.hardware_v2.lattice import AHSLatticeData
-from bloqade.codegen.hardware_v2.piecewise_linear import PiecewiseLinear
-from bloqade.codegen.hardware_v2.piecewise_constant import PiecewiseConstant
+from bloqade.compiler.codegen.hardware_v2.lattice import AHSLatticeData
+from bloqade.compiler.codegen.hardware_v2.piecewise_linear import PiecewiseLinear
+from bloqade.compiler.codegen.hardware_v2.piecewise_constant import PiecewiseConstant
 from beartype.typing import Optional, List
 from pydantic.dataclasses import dataclass
 from decimal import Decimal
