+++ src/bloqade/compiler/rewrite/common/assign_to_literal.py
