+++ src/bloqade/compiler/rewrite/common/canonicalize.py
