+++ src/bloqade/compiler/rewrite/common/flatten.py
