+++ src/bloqade/ir/control/traits/canonicalize.py
@@ -1,6 +1,6 @@
 class CanonicalizeTrait:
     @staticmethod
     def canonicalize(expr):
-        from bloqade.rewrite.common.canonicalize import Canonicalizer
+        from bloqade.compiler.rewrite.common.canonicalize import Canonicalizer
 
         return Canonicalizer().visit(expr)
