+++ src/bloqade/ir/scalar.py
@@ -181,7 +181,7 @@ class Scalar:
 
     @staticmethod
     def canonicalize(expr: "Scalar") -> "Scalar":
-        from bloqade.rewrite.common.canonicalize import Canonicalizer
+        from bloqade.compiler.rewrite.common.canonicalize import Canonicalizer
 
         return Canonicalizer().visit(expr)
