+++ tests/test_bravais.py
@@ -10,7 +10,7 @@ from bloqade.ir.location.bravais import (
 from bloqade.ir.location.bravais import Cell
 from bloqade import cast
 from math import sqrt
-from bloqade.codegen.common.json import BloqadeIRSerializer
+from bloqade.compiler.codegen.common.json import BloqadeIRSerializer
 
 
 def test_square():
