+++ tests/test_emulator_codegen.py
@@ -9,7 +9,7 @@ from bloqade.emulate.ir.emulator import (
     DetuningOperatorData,
     RabiOperatorType,
 )
-from bloqade.codegen.emulator_ir import (
+from bloqade.compiler.codegen.emulator_ir import (
     EmulatorProgramCodeGen,
     CompiledWaveform,
     LevelCoupling,
