+++ tests/test_hamiltonian_codegen.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 from itertools import combinations
 from bloqade.atom_arrangement import Chain
-from bloqade.codegen.emulator_ir import EmulatorProgramCodeGen
+from bloqade.compiler.codegen.emulator_ir import EmulatorProgramCodeGen
 from bloqade.emulate.codegen.hamiltonian import CompileCache, RydbergHamiltonianCodeGen
 from functools import reduce
 import numpy as np
